requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.4
email-validator==2.0.0
colorama==0.4.6
urllib3==2.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
MarkupSafe==2.1.5
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.5
Werkzeug==3.0.1
greenlet==3.0.3
SQLAlchemy==2.0.25
bcrypt==4.1.2
dnspython==2.6.1
cffi==1.16.0
pycparser==2.21
